LuisBotGit/Weather-Chatbot-Using-Luis-master/requirements.txt
botbuilder
adal==1.2.1
aiohttp==3.6.2
aiounittest==1.3.0
applicationinsights==0.11.9
APScheduler==3.6.3
async-timeout==3.0.1
attrs==19.3.0
azure-cognitiveservices-language-luis==0.2.0
azure-common==1.1.25
azure-cosmos==3.1.2
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
Babel==2.7.0
botbuilder-ai==4.8.0
botbuilder-applicationinsights==4.8.0
botbuilder-azure==4.8.0
botbuilder-core==4.8.0
botbuilder-dialogs==4.8.0
botbuilder-schema==4.8.0
botframework-connector==4.8.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorhash==1.0.2
configparser==4.0.2
cryptography==2.8
cycler==0.10.0
datedelta==1.3
Django==2.2.6
djangorestframework==3.10.3
emoji==0.5.4
Flask==1.1.1
Flask-Cors==3.0.8
Flask-MonitoringDashboard==3.0.6
geojson==2.5.0
grapheme==0.6.0
gunicorn==20.0.4
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.0
jsonpickle==1.2
kiwisolver==1.1.0
kneed==0.5.3
MarkupSafe==1.1.1
msal==1.1.0
msrest==0.6.10
msrestazure==0.6.3
multidict==4.7.5
multipledispatch==0.6.0
oauthlib==3.1.0
psutil==5.7.0
pycparser==2.20
PyJWT==1.5.3
pyowm==2.10.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.1
tzlocal==2.0.0
urllib3==1.25.8
Werkzeug==0.16.1
wincertstore==0.2
yarl==1.4.2
